ghl-frontend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.0.0
requests>=2.31.0
